backend/requirements.txt
Django==4.1.1
psycopg2-binary==2.9.3
djangorestframework==3.13.1
requests==2.28.1
model-bakery==1.7.0
django-cors-headers==3.13.0